refactor(ingestion): extract parser data and helpers from parser.py
parser.py was ~1,250 lines mixing the ASTParser class with two large,
independent concerns. Split those out, leaving parser.py as the parser
class itself (674 lines):

- language_configs.py: the LanguageConfig dataclass and the declarative
  LANGUAGE_CONFIGS table (per-language node-type/visibility/parent rules).
- parser_helpers.py: the stateless tree-sitter helper functions
  (_run_query, _build_qualified_name, _head_type_identifier, call-arg
  counting, enclosing-symbol lookup, etc.).

parser.py imports both back, so ASTParser, LANGUAGE_CONFIGS, LanguageConfig
and parse_file remain importable from the same path (verified against the
ingestion __init__ re-export and test_parser.py).

Pure relocation: the moved code is token-identical to the original (only
ruff-format line-wrapping differs); no logic changes. The existing
ingestion suite (578 passed, 2 xfailed) is the characterization guard.
